replace MutableHashMap
