+++ .changeset/weak-steaks-float.md
@@ -1,0 +1,5 @@
+---
+"effect": patch
+---
+
+optimize MutableHashMap
